Stand up the full stack for real: Temporal server, secrets, ports
Docker Hub / GHCR / registry access is unblocked now, so this actually starts
Temporal end to end for the first time — surfacing bugs no static check could
have caught:

- temporalio/admin-tools:1.29.0 and temporalio/ui:2.42.0 don't exist as tags;
  pin-digests.sh now resolves the real tags for this server version.
- admin-tools' image ENTRYPOINT is `tini -- sleep infinity` (meant for `docker
  exec`, not `docker run image CMD`), so bootstrap-temporal.sh's schema/
  namespace/search-attribute commands were silently never executing. Overridden
  with --entrypoint.
- Postgres's `-c=key=value` command args are invalid (postgres parses the `=`
  as part of a short option, giving an empty parameter name and refusing to
  start); split into separate `-c` / `key=value` list entries.
- The Temporal server image reads POSTGRES_PWD, not POSTGRES_PWD_FILE — there
  is no _FILE convention, so the mounted secret was silently ignored and the
  server always connected with an empty password. Read the secret into the
  process environment at container start instead.
- Compose (outside Swarm mode) bind-mounts each secret file as-is, with no
  uid/gid/mode translation, so a container's non-root user couldn't read a
  0600 host-owned secret. Secrets are now 0644 (the codec directory 0755),
  which is what any container user needs to read them at all.
- Ringpop rejects a DNS name for broadcastAddress; `TEMPORAL_BROADCAST_ADDRESS:
  temporal` doesn't work like TEMPORAL_ADDRESS does. Resolved to the
  container's real IP at start instead. BIND_ON_IP=0.0.0.0 added too — without
  it the frontend only listens on that same container IP, so nothing on
  127.0.0.1 answers (not the healthcheck, not the published port).
- SQL-backed visibility has a fixed pool of typed search-attribute columns (3
  Int, 3 Double, ...); the 8 attributes included 4 Int, one too many. AmountPence
  moved to Double — it's a search index, not the ledger.
- postgres-app/postgres-temporal sit only on `core` (internal: true), and
  Docker skips published-port setup entirely for a container whose sole
  network is internal — `127.0.0.1:5432` never actually answers, here or on a
  production VPS. bootstrap-app-db.sh now resolves the container's real
  core-network address instead.

With all of this, `pnpm verify:stack` runs 8/8 for real for the first time:
postgres-app, the bigint/append-only/codec/PII checks, the Temporal frontend,
search attributes, and — the check this script exists for — Workflow Update
end to end (T-10.8). The full `packages/db` integration suite (25 tests) also
passes against a live, migrated database. The Phase 1 gate is open.

+++ packages/temporal-common/src/task-queues.ts
@@ -33,6 +33,12 @@ export function assertTaskQueue(value: string): TaskQueue {
  * RUNTIME, not at deploy. The worker asserts their presence at startup rather
  * than discovering the omission during a draw.
  */
+// SQL-backed visibility has a fixed pool of typed columns per type (3 Int, 3
+// Double, ...) — Temporal rejects registration past the pool size, so a 4th
+// Int attribute does not fit. AmountPence uses Double instead: pence amounts
+// here are well inside a double's 53-bit exact-integer range, and this column
+// is a search index, not the ledger — postgres-app::bigint stays the only
+// source of truth for money (T-1.1).
 export const SEARCH_ATTRIBUTES = {
   DrawNumber: 'Int',
   DrawStatus: 'Keyword',
@@ -41,7 +47,7 @@ export const SEARCH_ATTRIBUTES = {
   StatementNumber: 'Int',
   TaskKind: 'Keyword',
   Blocked: 'Bool',
-  AmountPence: 'Int',
+  AmountPence: 'Double',
 } as const;
 
 export const NAMESPACE = process.env['TEMPORAL_NAMESPACE'] ?? 'qosfc-lottery';
